+++ CHANGELOG.md
@@ -8,6 +8,8 @@ REMOVED:
 
 FIXED:
 
+- Fix definition binding
+
 ## 2.5.1 - 2018-12-21
 
 CHANGED: